feat: use context to pass message id to image component in an attachment
